pyproject.toml
[project]
name = "ai-hedge-fund"
version = "0.1.0"
description = "AI-powered hedge fund using multi-agent LLM system for US stock trading"
authors = [{name = "DanisHack"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    # Agent framework
    "langchain>=0.3.7",
    "langgraph>=0.2.56",
    "langchain-openai>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "langchain-groq>=0.2.0",
    "langchain-google-genai>=2.0.0",
    # Market data
    "polygon-api-client>=1.14.0,<1.16.0",
    # Broker
    "alpaca-py>=0.31.0",
    # Data & analysis
    "pandas>=2.1.0",
    "numpy>=1.24.0",
    # Utilities
    "python-dotenv>=1.0.0",
    "pydantic>=2.4.2",
    "rich>=13.9.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.6.0",
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
target-version = "py311"
